requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Django==5.2.7
emoji==2.15.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
image==1.5.33
joblib==1.5.2
nltk==3.9.2
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.3
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
